+++ CHANGELOG.md
@@ -1,8 +1,20 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
+The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+
+---
+
+## [0.2.0] - 2025-12-22
+### Changed
+- Migrated from Flask to FastAPI
+- Replaced SQLAlchemy with Pydantic models and in-memory storage
+- Modernized packaging to use pyproject.toml exclusively
+
+### Removed
+- SQLite database dependency
+- Legacy setup.py, setup.cfg, and MANIFEST.in files
 
 ---
 
@@ -32,11 +44,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.0.1] - 2017-10-19
 ### Added
-- Basic project setup (readme, changelog, CI, gitignore...) 
+- Basic project setup (readme, changelog, CI, gitignore...)
 - Flask application first layout
 - Functional tests
 - Resource Model and GET REST api
 - Resource Lock and Unlock API
 - A (very) basic front end to display resources
 - Resource descriptor yaml file handling at startup
-